The StoryEditor now keeps its list of functions in sync with its StoryMap, adding and removing functions as they are added to and removed from the StoryMap.
The TextArea for each function is stored with the StoryCard, so it persists
even as the function is removed from and added back to the StoryEditor.

StoryBase: had to make background protected, because StoryMap needs
to access it. Alternatively could have added an accessor method,
or could add some drag-drop functionality into StoryBase. Have
StoryBase figure out whether it is something from elsewhere
dropped onto this droppable, something from this droppable dropped
onto something else, or something from this droppable dropped onto
this droppable again, and call one of three methods in StoryMap or
StoryCards for each case.

StoryCard: associated a new FunctionEditor instance field with StoryCard, a
StoryCard constructs its FunctionEditor when it constructs itself.

FunctionEditor: new class. Wraps a JPanel that it gives a FlowLayout and
attached a JLabel and a JTextArea to, forming a "paragraph" for the editing of
one function in a story.

StoryEditor: new class. Holds a reference to a StoryMap that it keeps itself in
sync with. Has a JPanel wrapped in a JScrollPane and with a BoxLayout, that it
adds and removes FunctionEditors to/from. Implements the Receiver interface to
get notified when the StoryMap has changed.

StoryMap: broadcast message "StoryMap changed" whenever a card is added to or
removed from the story map. Be aware of when a card is dragged from this story
map and dropped back onto this same story map, in that case don't remove the
card from the story map.

Swing: changed into a Swing GUI for all the new stuff, and embedded the Piccolo
canvas into the Swing GUI, moving code over from Main that initialises the
Piccolo canvas.

	new file:   FunctionEditor.java
	modified:   StoryBase.java
	modified:   StoryCard.java
	modified:   StoryEditor.java
	modified:   StoryMap.java
	modified:   Swing.java
	modified:   TODO.txt