modified:   StoryEditor.java 	modified:   StoryMap.java 	modified:   Swing.java
Refactoring of StoryMap and StoryEditor so that StoryMap holds a reference to a
StoryEditor and not the other way round, and they no longer need to use
Messager to communicate.

This now gets the StoryEditor updating as the StoryMap changes, including if
the StoryMap is restored. Seems to work, though needs proper testing, and there
seems to be a couple of bugs with the JScrollPane not updating itself right
away when FunctionEditors are removed.

StoryEditor:

- No longer implements Receiver, removed receive method, no
  longer accepts any messages from Messager.
- Removed StoryMap field.
+ Renamed updateFunctions to update, made it public, and gave
  it an ArrayList<StoryCard> as an argument. This is called by
  StoryMap whenever the functions need to be updated.
+ Added public method focus(StoryCard s) so that the StoryMap
  can tell the StoryEditor to focus a particular function when
  a StoryCard is focused in the StoryMap.

StoryMap:

+ Added a new argument, StoryEditor, to the constructor.
+ Replaced sending messages via Messager with calling methods
  on the StoryEditor directly.

Swing: updated to reflect these changes.
Random bugfix: StoryMap now has 33 placeholders not 31, so they fill the entire
background rectangle.