When a draggable is dropped, search the entire pick path for a node that is not the dragged node and that has a "Droppable" attribute, instead of just looking at the first node in the pick path that is not the dragged node.
This allows you to drop a story card anywhere on a story map, e.g. on top of
another story card, and still have the drop be picked up by the story map.

The code in the event listener in draggable is getting pretty hairy, could do
with being broken up.

	modified:   ../DragAndDrop/Draggable.java