Make story cards remain their normal (non-scaled) size while being dragged.
This is a bit messy. Had to hack the drag-drop so that a draggable uses
Messager to send a message when the user starts to drag it. (Draggable uses its
own set of notification interfaces to send messages usually, but perhaps it
would be better to use Messager for everything anyway.)

Had to make the Message class public so that Draggable (in a different package)
can use it.

StoryCard receives the new drag-started message and also receives drop messages
from its draggable, and uses these notifications to disable scaling while being
dragged.

Also cut out some lines of old unused code from StoryCard.

	modified:   DragAndDrop/Draggable.java
	modified:   storymaps/Messager.java
	modified:   storymaps/StoryCard.java