Integrated the taskbar implemented in Piccolo from ZoomDesk. This involves creating several new classes which I placed in a new storymaps.ui package.
It could still do with some hacking. The height of the bar (and the buttons
and the font sizes) should maybe be dependent on the size of the canvas.

The popup dialogues should probably use PText to get text that wraps
automatically. A dialog could have three PText nodes on it: title, text, and
the instruction for closing the dialog (which is the same on all dialogs). That
would still allow you to use three different text sizes. That way you could
probably implement just one class, instead of having a different subclass for
each dialog.

New files in storymaps/ui:

+ Bar.java: the piccolo task bar
+ Colours.java: a file for holding all the colour specifications, good idea
from ZoomDesk, should use this everywhere.
+ Fonts.java
+ Strokes.java
+ HelpButton.java
+ Popup.java: base class for popup messages
+ WelcomePopup.java

Main.java:

+ new field Bar bar.
+ new method createBar()
- commented out method makeToolBar that adds the swing toolbar
+ restrict the bounds of the piccolo camera so that they don't overlap the bar.
(Prevents objects from going under the bar.)
+ Move initial positioning of piccolo camera to after the bar has been created.
+ Resize the piccolo bar when Swing detects a resize of the window
+ Added methods centre, getCamera, getCanvas, disableHotkeys, enableHotkeys, to
make the new classes work.