Updating the text for Propp's functions with new revised version.
This completely removes XML from the application, the functions.xml file is now
replaced with a functions.json file that is parsed with org.json.

- Removed XStream from lib.
- Removed XMLHandler.
- Removed FunctionConverter.
- Removed functions.xml.

+ Added json folder containing the source code (java files) for org.json.
+ Added functions.json.

Function:

* Refactored the class a bit so that functions have the fields number, name,
description and instruction (plus the fields for the image).

* Added code to parse the functions.json file using org.json and construct
Function objects.

FunctionEditor:

* Refactored it a bit so that a FunctionEditor has fields name, image and
instructions. instructions is a JEditorPane displaying the text of the
description and instructions field from the function.

StoryCardBase:

* title_node displays the name of the function, while description_node displays
the description.
* Also removed some unnecessary publics.

StoryCardDemo and TemplateHandler: minor fixes due to changed names of fields
in Function.