Fix a bug where story cards would sometimes highlight and unhighlight (scale up and down) while being dragged.
StoryCard don't try to keep track of whether it's being dragged or not, that's
a horrible design, instead have Draggable keep track of that and have StoryCard
call draggable.isDragging().

You can still sometimes highlight the story cards that you aren't dragging by
moving over them when you are dragging one. To fix that there needs to be a
global isAnythinggBeingDragged() somewhere.