Polishing the appearance of the application by implementing Ray's mockup with new colours and layout. Also added drop shadows to story cards. Things look a lot nicer now, but the code needs a lot of refactoring.
The main changes are in Application, to cause the new green and grey rectangles
to fill the proportions of the canvas, in StoryBase to implement the new
centre-justified rows layout (this should have been implemented as a new layout
node), in StoryEditor to add a new toolbar with the write story and sort
buttons, and in StoryCardBase to add the drop shadows.

- No longer use a VerticalLayoutNode as the home node. Just use a PNode.
Positioning and size of the StoryCards and StoryMap objects is done manually to
fit the initial proportions of the canvas. (Although this could easily have
been done using a layout node with margin 0.)

- No longer has a WriteStoryButton object. This has been replaced with a Swing
button on the StoryEditor.

- The second_home node is no longer needed.

+ Set the background colour of the canvas to BLACK, not DARK_GRAY.

+ StoryCards and StoryMap no longer have text titles.

+ Rewrote how the camera is repositioned when the story editor is collapsed and
uncollapsed. We now remove the StoryCards object from the scene (probably
unnecessary) and put it back, and schedule a camera reposition for 100secs time.

Rewrote to use Ray's centre-justified layout. Should have implemented this as a
layout node instead. Never mind.

- No longer have a vnode as the root node.
- No longer have a PText node as the title
- Got rid of the GridLayoutNode for the grid.
- Got rid of the overlay PNode.

The way it works now is by adding both background-grid and overlay nodes to the
background rectangle node. References to the background grid nodes are kept in a
list so we know which ones they are. This should all be redone with reusable
layout nodes and with StoryCard objects and can hold other StoryCard objects on
top of them.

+ New method layoutGrid implements a centre-justified layout within the bounds
of the rectangle. This is what should go in a CentreJustifiedRows layout node.

- Deleted StoryBaseDemo, couldn't be bothered to update it to deal with these
changes.

of the sort button.

+ Added a new toolbar that remains visible at the bottom of the application
when the story editor is collapsed. This toolbar contains the 'Write your
Story!' and 'Sort' buttons, which were previously implemented within the
Piccolo scene.

+ Added drop-shadows. There is now a root node to which the card node and
shadow node are attached.

+ Added newRectangle method to create the story card rectangle shapes. Exported
for other classes (placeholder) to use.

+ Replace calls to getBackground() with calls to getNode().

+ Set shadow node not visible.

+ Call StoryCard.newRectangle() rather than duplicating code.

+ Call setVisible(false) instead of setTransparency(0), probably does the same
thing, but this might be faster.