Adjusting the layout (and all the hardcoded figures) to reduce the amount of space in the story map to 18 story cards in two rows.
Also adjusted the implementation of StoryBase to allow for top and left margins
so that story cards don't go off screen when they scale up. Not sure what I
think about this yet, it might make the text too small.

* Adjusted the size of story cards, relative size of the StoryCards background
and the StoryMap background, how far story cards scale up on mouse over, and
size of left and top margins of StoryCards and StoryMap.

- StoryCards and StoryMap no longer take `margin` as a parameter (and no longer
save it in the saved files!), this was madness. Color should not be a param
either.

+ Added margin_left and margin_top params to StoryBase. StoryCards and StoryMap
have settings for these params in class constants.

+ StoryBase goes back to having an overlay node so that story cards can be offset relative to background.