Use the logging API the way it's supposed to be used.
This fixes a NullPointerException that would cause the application to crash if
some code run during Application's constructor attempted to log a message, e.g.
because an image file could not be found.

Code should not need to access the singleton instance of Application to log
messages. Instead each class accesses its own Logger instance via Java's
getLogger method. A FileHandler is attached to the root Logger so that log
messages are saved to file.

This commit also includes lot of formatting changes to Application created by
NetBeans' autoformat function.