Added internal printing and preview and reworked export-to-html a little. Print, preview and save-as-html all work by converting the storymap to HTML, using the same templates, but save-as-html additionally exports the storycard images to a folder. The card backgrounds don't show up when previewing or when printing from within Java, but that won't be a problem anymore when branch ray is merged.
- Change the HTML templates. They have no HTML header because Swing's HTML
loader can't handle it and they use very little CSS because any CSS has to be
both supported by Swing and IE 6. Swing has _very_ little CSS support. Tables
are used for the layout, and CSS is used only to set fonts and align things
within tables. It doesn't look as beautiful, but it should be much more
reliable.

+ Added class PreviewDialog. Implements a popup preview before printing or
exporting the story. The print and save-as-html code is shared between this
class and TemplateHandler. The design is a little odd, really a PreviewDialog
should belong to a StoryMap, and the printing and saving-as-html should be
handled by StoryMap which should have its own TemplateHandler instance (if
TemplateHandler should exist at all). Not time to worry about refactoring now.

+ Also added print and preview icons.